requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
channels-redis
pillow
drf-yasg
